make transformers own folder
