+++ scylla-server-rust/tests/data_service_test.rs
@@ -4,9 +4,9 @@ mod test_utils;
 use prisma_client_rust::QueryError;
 use protobuf::SpecialFields;
 use scylla_server_rust::{
-    controllers::data_controller::PublicData,
     serverdata::ServerData,
     services::{data_service, data_type_service, node_service, run_service},
+    transformers::data_transformer::PublicData,
 };
 use test_utils::cleanup_and_prepare;
